Rename algolia to typesense / instantsearch
