+++ assets/scripts/components/async-loading.js
@@ -7,7 +7,7 @@ import { initializeGroupedListings } from './grouped-item-listings';
 import {updateMainContentAnchors, reloadWistiaVidScripts, gtag, getCookieByName } from '../helpers/helpers';
 import configDocs from '../config/config-docs';
 import { redirectCodeLang, addCodeTabEventListeners, addCodeBlockVisibilityToggleEventListeners, activateCodeLangNav, toggleMultiCodeLangNav } from './code-languages'; // eslint-disable-line import/no-cycle
-import { loadInstantSearch } from './algolia';
+import { loadInstantSearch } from './instantsearch';
 
 const { env } = document.documentElement.dataset;
 const { gaTag } = configDocs[env];
