+++ assets/scripts/components/instantsearch/getHitData.js
