+++ assets/scripts/components/instantsearch/searchbarHits.js
