+++ assets/scripts/components/instantsearch/searchpageHits.js
