+++ assets/scripts/datadog-docs.js
@@ -7,7 +7,7 @@ import { initializeGroupedListings } from './components/grouped-item-listings';
 import { updateTOC, buildTOCMap, onScroll, closeMobileTOC } from './components/table-of-contents';
 import initCodeTabs from './components/codetabs';
 import { loadPage } from './components/async-loading';
-import { loadInstantSearch } from './components/algolia';
+import { loadInstantSearch } from './components/instantsearch';
 import { setMobileNav, closeMobileNav } from './components/mobile-nav';
 
 const { env } = document.documentElement.dataset;
@@ -81,7 +81,7 @@ const doOnLoad = () => {
         });
     }
 
-    // Load algolia instant search for the first time
+    // Load instant search for the first time
     loadInstantSearch((asyncLoad = false));
 
     if (!bodyClassContains('api')) {
