+++ assets/styles/_instantsearch.scss
